Refactor classes into separate files and only import required based on config
